requirements.txt
streamlit>=1.25.0

PyMuPDF==1.22.3

sentence-transformers>=2.6.0
sentencepiece==0.1.99
torch>=1.13.0
transformers>=4.30.0

huggingface_hub>=0.16.4

google-generativeai>=0.7.0
google-ai-generativelanguage<0.7.0,>=0.6.6
google-cloud-aiplatform>=1.36.0

chromadb==0.4.15

tqdm
numpy>=1.24.3

librosa>=0.9.0
